requirements.txt
alembic==1.15.1
arrow==1.3.0
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
Cython==3.0.12
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-DebugToolbar==0.16.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2025.3.0
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
psycopg2==2.9.10
pybind11==2.13.6
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==76.1.0
six==1.17.0
SQLAlchemy==2.0.39
sympy==1.13.1
text-unidecode==1.3
torch==2.6.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3
Werkzeug==3.1.3
WTForms==3.2.1
flasgger==0.9.7.1
